app-consumer/app_code/uai3071390-genai-services-demand-generation-usecase/backend/agents/question-answer-agent/pyproject.toml
# ==============================================================================
# sdg-qna-agent — uv workspace member
# Strands SDK + FastAPI agentic assistant
# Port: 8005
# Full Strands pattern: Agent() loop + LiteLLMModel + S3SessionManager
# ==============================================================================

[project]
name = "sdg-qna-agent"
version = "0.1.0"
description = "Ad Hoc Q&A Agent — Strands SDK + LiteLLM (agentic)"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "boto3>=1.34.0",
    "PyJWT>=2.8.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "sdg-commons",
    # Security pin: litellm 1.82.7 and 1.82.8 were compromised (PYSEC-2026-2 / MAL-2026-2144).
    "litellm==1.82.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
sdg-commons = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/question_answer"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
